VERSION: Move version.cpp from libcommon.la into its own libversion.la
