+++ configure.ac
@@ -262,6 +262,7 @@ AC_CONFIG_FILES([utf8cpp/Makefile])
 AC_CONFIG_FILES([lua/Makefile])
 AC_CONFIG_FILES([toluapp/Makefile])
 AC_CONFIG_FILES([gitstamp/Makefile])
+AC_CONFIG_FILES([src/version/Makefile])
 AC_CONFIG_FILES([src/common/Makefile])
 AC_CONFIG_FILES([src/graphics/Makefile])
 AC_CONFIG_FILES([src/graphics/images/Makefile])
