+++ src/common/Makefile.am
@@ -21,22 +21,6 @@
 
 include $(top_srcdir)/Makefile.common
 
-# Pass the git version information to the compiler
-
-VER_REVDESC = $(shell cat $(top_builddir)/gitstamp/gitstamp)
-VERSION_FLAGS = $(shell if test x$(VER_REVDESC) != "x"; then echo "-DXOREOS_REVDESC=\\\"$(VER_REVDESC)\\\""; fi)
-
-VER_REV = $(shell cat $(top_builddir)/gitstamp/gitstamp | cut -d '+' -f 2-)
-VERSION_FLAGS += $(shell if test x$(VER_REV) != "x"; then echo "-DXOREOS_REV=\\\"$(VER_REV)\\\""; fi)
-
-VER_DATE = $(shell date -u '+%Y-%m-%dT%H:%M:%SZ')
-VERSION_FLAGS += $(shell if test x$(VER_DATE) != "x"; then echo "-DXOREOS_BUILDDATE=\\\"$(VER_DATE)\\\""; fi)
-
-AM_CXXFLAGS += $(VERSION_FLAGS)
-
-# Rebuild the version.lo when the git version changed
-version.lo: $(top_builddir)/gitstamp/gitstamp
-
 noinst_LTLIBRARIES = libcommon.la
 
 noinst_HEADERS = \
@@ -44,7 +28,6 @@ noinst_HEADERS = \
                  noreturn.h \
                  types.h \
                  endianness.h \
-                 version.h \
                  noncopyable.h \
                  singleton.h \
                  maths.h \
@@ -101,7 +84,6 @@ noinst_HEADERS = \
                  $(EMPTY)
 
 libcommon_la_SOURCES = \
-                       version.cpp \
                        maths.cpp \
                        sinetables.cpp \
                        cosinetables.cpp \
