+++ src/engines/gamethread.cpp
@@ -24,7 +24,8 @@
 
 #include <cassert>
 
-#include "src/common/version.h"
+#include "src/version/version.h"
+
 #include "src/common/util.h"
 #include "src/common/error.h"
 #include "src/common/ustring.h"
